Add Card Works, DeckDetail attached to state instead of props.
Add card updates the deck and re-navigates to the deck detail. Tried passing the item through the add card to deckdetail, but needed to refresh to see the increase in card count. Therefore attached the Deck Detail directly to the state to allow it to fetch latest detail about the card. Now it only recieves the deck title from other components but fetches the card on it's own.